+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 * added DeferredValue
 * satella.coding.Context are considered experimental
+* added a way to register an object to be cleaned up via MemoryPressureManager